+++ pylabrobot/liquid_handling/backends/hamilton/vantage.py
@@ -4,13 +4,8 @@ import re
 import sys
 from typing import Dict, List, Optional, Sequence, Union, cast
 
-from pylabrobot.liquid_handling.backends.hamilton.base import (
-  HamiltonLiquidHandler,
-)
-from pylabrobot.liquid_handling.liquid_classes.hamilton import (
-  HamiltonLiquidClass,
-  get_vantage_liquid_class,
-)
+from pylabrobot.liquid_handling.backends.hamilton.base import HamiltonLiquidHandler
+from pylabrobot.liquid_handling.liquid_classes.hamilton import HamiltonLiquidClass
 from pylabrobot.liquid_handling.standard import (
   Pickup,
   PickupTipRack,
@@ -36,6 +31,8 @@ from pylabrobot.resources.ml_star import (
   TipPickupMethod,
   TipSize,
 )
+from pylabrobot.resources import Coordinate, Resource, TipRack, Well
+from pylabrobot.resources.ml_star import HamiltonTip, TipPickupMethod, TipSize
 
 
 if sys.version_info >= (3, 8):
@@ -566,8 +563,6 @@ class Vantage(HamiltonLiquidHandler):
     self,
     ops: List[Aspiration],
     use_channels: List[int],
-    jet: Optional[List[bool]] = None,
-    blow_out: Optional[List[bool]] = None,
     hlcs: Optional[List[Optional[HamiltonLiquidClass]]] = None,
     type_of_aspiration: Optional[List[int]] = None,
     minimal_traverse_height_at_begin_of_command: Optional[List[float]] = None,
@@ -614,51 +609,20 @@ class Vantage(HamiltonLiquidHandler):
       blow_out: Whether to search for a "blow out" liquid class. This is only used on dispense.
         Note that in the VENUS liquid editor, the term "empty" is used for this, but in the firmware
         documentation, "empty" is used for a different mode (dm4).
-      hlcs: The Hamiltonian liquid classes to use. If `None`, the liquid classes will be
-        determined automatically based on the tip and liquid used.
     """
 
-    x_positions, y_positions, channels_involved = self._ops_to_fw_positions(ops, use_channels)
+    if hlcs is not None:
+      raise NotImplementedError("hlcs is deprecated")
 
-    if jet is None:
-      jet = [False] * len(ops)
-    if blow_out is None:
-      blow_out = [False] * len(ops)
-
-    if hlcs is None:
-      hlcs = []
-      for j, bo, op in zip(jet, blow_out, ops):
-        liquid = Liquid.WATER  # default to WATER
-        # [-1][0]: get last liquid in well, [0] is indexing into the tuple
-        if len(op.liquids) > 0 and op.liquids[-1][0] is not None:
-          liquid = op.liquids[-1][0]
-        hlcs.append(
-          get_vantage_liquid_class(
-            tip_volume=op.tip.maximal_volume,
-            is_core=False,
-            is_tip=True,
-            has_filter=op.tip.has_filter,
-            liquid=liquid,
-            jet=j,
-            blow_out=bo,
-          )
-        )
+    x_positions, y_positions, channels_involved = \
+      self._ops_to_fw_positions(ops, use_channels)
 
     self._assert_valid_resources([op.resource for op in ops])
 
-    # correct volumes using the liquid class
-    volumes = [
-      hlc.compute_corrected_volume(op.volume) if hlc is not None else op.volume
-      for op, hlc in zip(ops, hlcs)
-    ]
-
-    well_bottoms = [
-      op.resource.get_absolute_location().z + op.offset.z + op.resource.material_z_thickness
-      for op in ops
-    ]
-    liquid_surfaces_no_lld = liquid_surface_at_function_without_lld or [
-      wb + (op.liquid_height or 0) for wb, op in zip(well_bottoms, ops)
-    ]
+    well_bottoms = [op.resource.get_absolute_location().z + op.offset.z + \
+                    op.resource.material_z_thickness for op in ops]
+    liquid_surfaces_no_lld = liquid_surface_at_function_without_lld or [wb + (op.liquid_height or 0)
+                              for wb, op in zip(well_bottoms, ops)]
     # -1 compared to STAR?
     lld_search_heights = lld_search_height or [
       wb
@@ -667,14 +631,8 @@ class Vantage(HamiltonLiquidHandler):
       for wb, op in zip(well_bottoms, ops)
     ]
 
-    flow_rates = [
-      op.flow_rate or (hlc.aspiration_flow_rate if hlc is not None else 100)
-      for op, hlc in zip(ops, hlcs)
-    ]
-    blow_out_air_volumes = [
-      (op.blow_out_air_volume or (hlc.dispense_blow_out_volume if hlc is not None else 0))
-      for op, hlc in zip(ops, hlcs)
-    ]
+    flow_rates = [op.flow_rate or 100 for op in ops]
+    blow_out_air_volumes = [op.blow_out_air_volume or 0 for op in ops]
 
     return await self.pip_aspirate(
       x_position=x_positions,
@@ -705,39 +663,31 @@ class Vantage(HamiltonLiquidHandler):
         round(t2sr * 10) for t2sr in tube_2nd_section_ratio or [0] * len(ops)
       ],
       minimum_height=[round(wb * 10) for wb in minimum_height or well_bottoms],
-      immersion_depth=[round(id_ * 10) for id_ in immersion_depth or [0] * len(ops)],
-      surface_following_distance=[
-        round(sfd * 10) for sfd in surface_following_distance or [0] * len(ops)
-      ],
-      aspiration_volume=[round(vol * 100) for vol in volumes],
+      immersion_depth=[round(id_*10) for id_ in immersion_depth or [0]*len(ops)],
+      surface_following_distance=[round(sfd*10) for sfd in surface_following_distance or
+                                  [0]*len(ops)],
+      aspiration_volume=[round(op.volume*100) for op in ops],
       aspiration_speed=[round(fr * 10) for fr in flow_rates],
-      transport_air_volume=[
-        round(tav * 10)
-        for tav in transport_air_volume
-        or [hlc.aspiration_air_transport_volume if hlc is not None else 0 for hlc in hlcs]
-      ],
-      blow_out_air_volume=[round(bav * 100) for bav in blow_out_air_volumes],
-      pre_wetting_volume=[round(pwv * 100) for pwv in pre_wetting_volume or [0] * len(ops)],
-      lld_mode=lld_mode or [0] * len(ops),
-      lld_sensitivity=lld_sensitivity or [4] * len(ops),
-      pressure_lld_sensitivity=pressure_lld_sensitivity or [4] * len(ops),
-      aspirate_position_above_z_touch_off=[
-        round(apz * 10) for apz in aspirate_position_above_z_touch_off or [0.5] * len(ops)
-      ],
-      swap_speed=[round(ss * 10) for ss in swap_speed or [2] * len(ops)],
-      settling_time=[round(st * 10) for st in settling_time or [1] * len(ops)],
-      mix_volume=[round(mv * 100) for mv in mix_volume or [0] * len(ops)],
-      mix_cycles=mix_cycles or [0] * len(ops),
-      mix_position_in_z_direction_from_liquid_surface=[
-        round(mp) for mp in mix_position_in_z_direction_from_liquid_surface or [0] * len(ops)
-      ],
-      mix_speed=[round(ms * 10) for ms in mix_speed or [250] * len(ops)],
-      surface_following_distance_during_mixing=[
-        round(sfdm * 10) for sfdm in surface_following_distance_during_mixing or [0] * len(ops)
-      ],
-      TODO_DA_5=TODO_DA_5 or [0] * len(ops),
-      capacitive_mad_supervision_on_off=capacitive_mad_supervision_on_off or [0] * len(ops),
-      pressure_mad_supervision_on_off=pressure_mad_supervision_on_off or [0] * len(ops),
+      transport_air_volume=[round(tav*10) for tav in transport_air_volume or [0]*len(ops)],
+      blow_out_air_volume=[round(bav*100) for bav in blow_out_air_volumes],
+      pre_wetting_volume=[round(pwv*100) for pwv in pre_wetting_volume or [0]*len(ops)],
+      lld_mode=lld_mode or [0]*len(ops),
+      lld_sensitivity=lld_sensitivity or [4]*len(ops),
+      pressure_lld_sensitivity=pressure_lld_sensitivity or [4]*len(ops),
+      aspirate_position_above_z_touch_off=
+        [round(apz*10) for apz in aspirate_position_above_z_touch_off or [0.5]*len(ops)],
+      swap_speed=[round(ss*10) for ss in swap_speed or [2]*len(ops)],
+      settling_time=[round(st*10) for st in settling_time or [1]*len(ops)],
+      mix_volume=[round(mv*100) for mv in mix_volume or [0]*len(ops)],
+      mix_cycles=mix_cycles or [0]*len(ops),
+      mix_position_in_z_direction_from_liquid_surface=
+        [round(mp) for mp in mix_position_in_z_direction_from_liquid_surface or [0]*len(ops)],
+      mix_speed=[round(ms*10) for ms in mix_speed or [250]*len(ops)],
+      surface_following_distance_during_mixing=
+        [round(sfdm*10) for sfdm in surface_following_distance_during_mixing or [0]*len(ops)],
+      TODO_DA_5=TODO_DA_5 or [0]*len(ops),
+      capacitive_mad_supervision_on_off=capacitive_mad_supervision_on_off or [0]*len(ops),
+      pressure_mad_supervision_on_off=pressure_mad_supervision_on_off or [0]*len(ops),
       tadm_algorithm_on_off=tadm_algorithm_on_off or 0,
       limit_curve_index=limit_curve_index or [0] * len(ops),
       recording_mode=recording_mode or 0,
@@ -791,8 +741,6 @@ class Vantage(HamiltonLiquidHandler):
     Args:
       ops: The aspiration operations.
       use_channels: The channels to use.
-      hlcs: The Hamiltonian liquid classes to use. If `None`, the liquid classes will be
-        determined automatically based on the tip and liquid used.
 
       jet: Whether to use jetting for each dispense. Defaults to `False` for all. Used for
         determining the dispense mode. True for dispense mode 0 or 1.
@@ -804,7 +752,11 @@ class Vantage(HamiltonLiquidHandler):
         documentation. Dispense mode 4.
     """
 
-    x_positions, y_positions, channels_involved = self._ops_to_fw_positions(ops, use_channels)
+    if hlcs is not None:
+      raise NotImplementedError("hlcs is deprecated")
+
+    x_positions, y_positions, channels_involved = \
+      self._ops_to_fw_positions(ops, use_channels)
 
     if jet is None:
       jet = [False] * len(ops)
@@ -813,38 +765,13 @@ class Vantage(HamiltonLiquidHandler):
     if blow_out is None:
       blow_out = [False] * len(ops)
 
-    if hlcs is None:
-      hlcs = []
-      for j, bo, op in zip(jet, blow_out, ops):
-        liquid = Liquid.WATER  # default to WATER
-        # [-1][0]: get last liquid in tip, [0] is indexing into the tuple
-        if len(op.liquids) > 0 and op.liquids[-1][0] is not None:
-          liquid = op.liquids[-1][0]
-        hlcs.append(
-          get_vantage_liquid_class(
-            tip_volume=op.tip.maximal_volume,
-            is_core=False,
-            is_tip=True,
-            has_filter=op.tip.has_filter,
-            liquid=liquid,
-            jet=j,
-            blow_out=bo,
-          )
-        )
-
     self._assert_valid_resources([op.resource for op in ops])
 
-    # correct volumes using the liquid class
-    volumes = [
-      hlc.compute_corrected_volume(op.volume) if hlc is not None else op.volume
-      for op, hlc in zip(ops, hlcs)
-    ]
+    well_bottoms = [op.resource.get_absolute_location().z + op.offset.z + \
+                    op.resource.material_z_thickness for op in ops]
+    liquid_surfaces_no_lld = [wb + (op.liquid_height or 0)
+                              for wb, op in zip(well_bottoms, ops)]
 
-    well_bottoms = [
-      op.resource.get_absolute_location().z + op.offset.z + op.resource.material_z_thickness
-      for op in ops
-    ]
-    liquid_surfaces_no_lld = [wb + (op.liquid_height or 0) for wb, op in zip(well_bottoms, ops)]
     # -1 compared to STAR?
     lld_search_heights = lld_search_height or [
       wb
@@ -853,15 +780,8 @@ class Vantage(HamiltonLiquidHandler):
       for wb, op in zip(well_bottoms, ops)
     ]
 
-    flow_rates = [
-      op.flow_rate or (hlc.dispense_flow_rate if hlc is not None else 100)
-      for op, hlc in zip(ops, hlcs)
-    ]
-
-    blow_out_air_volumes = [
-      (op.blow_out_air_volume or (hlc.dispense_blow_out_volume if hlc is not None else 0))
-      for op, hlc in zip(ops, hlcs)
-    ]
+    flow_rates = [op.flow_rate or 100 for op in ops]
+    blow_out_air_volumes = [op.blow_out_air_volume or 0 for op in ops]
 
     type_of_dispensing_mode = type_of_dispensing_mode or [
       _get_dispense_mode(jet=jet[i], empty=empty[i], blow_out=blow_out[i]) for i in range(len(ops))
@@ -872,62 +792,45 @@ class Vantage(HamiltonLiquidHandler):
       y_position=y_positions,
       tip_pattern=channels_involved,
       type_of_dispensing_mode=type_of_dispensing_mode,
-      minimum_height=[round(wb * 10) for wb in minimum_height or well_bottoms],
-      lld_search_height=[round(sh * 10) for sh in lld_search_heights],
-      liquid_surface_at_function_without_lld=[round(ls * 10) for ls in liquid_surfaces_no_lld],
-      pull_out_distance_to_take_transport_air_in_function_without_lld=[
-        round(pod * 10)
-        for pod in pull_out_distance_to_take_transport_air_in_function_without_lld
-        or [5.0] * len(ops)
-      ],
-      immersion_depth=[round(id * 10) for id in immersion_depth or [0] * len(ops)],
-      surface_following_distance=[
-        round(sfd * 10) for sfd in surface_following_distance or [2.1] * len(ops)
-      ],
-      tube_2nd_section_height_measured_from_zm=[
-        round(t2sh * 10) for t2sh in tube_2nd_section_height_measured_from_zm or [0] * len(ops)
-      ],
-      tube_2nd_section_ratio=[
-        round(t2sr * 10) for t2sr in tube_2nd_section_ratio or [0] * len(ops)
-      ],
-      minimal_traverse_height_at_begin_of_command=[
-        round(mth * 10)
-        for mth in minimal_traverse_height_at_begin_of_command
-        or [self._traversal_height] * len(ops)
-      ],
-      minimal_height_at_command_end=[
-        round(mh * 10)
-        for mh in minimal_height_at_command_end or [self._traversal_height] * len(ops)
-      ],
-      dispense_volume=[round(vol * 100) for vol in volumes],
-      dispense_speed=[round(fr * 10) for fr in flow_rates],
-      cut_off_speed=[round(cs * 10) for cs in cut_off_speed or [250] * len(ops)],
-      stop_back_volume=[round(sbv * 100) for sbv in stop_back_volume or [0] * len(ops)],
-      transport_air_volume=[
-        round(tav * 10)
-        for tav in transport_air_volume
-        or [hlc.dispense_air_transport_volume if hlc is not None else 0 for hlc in hlcs]
-      ],
-      blow_out_air_volume=[round(boav * 100) for boav in blow_out_air_volumes],
-      lld_mode=lld_mode or [0] * len(ops),
-      side_touch_off_distance=round(side_touch_off_distance * 10),
-      dispense_position_above_z_touch_off=[
-        round(dpz * 10) for dpz in dispense_position_above_z_touch_off or [0.5] * len(ops)
-      ],
-      lld_sensitivity=lld_sensitivity or [1] * len(ops),
-      pressure_lld_sensitivity=pressure_lld_sensitivity or [1] * len(ops),
-      swap_speed=[round(ss * 10) for ss in swap_speed or [1] * len(ops)],
-      settling_time=[round(st * 10) for st in settling_time or [0] * len(ops)],
-      mix_volume=[round(mv * 100) for mv in mix_volume or [0] * len(ops)],
-      mix_cycles=mix_cycles or [0] * len(ops),
-      mix_position_in_z_direction_from_liquid_surface=[
-        round(mp) for mp in mix_position_in_z_direction_from_liquid_surface or [0] * len(ops)
-      ],
-      mix_speed=[round(ms * 10) for ms in mix_speed or [1] * len(ops)],
-      surface_following_distance_during_mixing=[
-        round(sfdm * 10) for sfdm in surface_following_distance_during_mixing or [0] * len(ops)
-      ],
-      TODO_DD_2=TODO_DD_2 or [0] * len(ops),
+      minimum_height=[round(wb*10) for wb in minimum_height or well_bottoms],
+      lld_search_height=[round(sh*10) for sh in lld_search_heights],
+      liquid_surface_at_function_without_lld=[round(ls*10) for ls in liquid_surfaces_no_lld],
+      pull_out_distance_to_take_transport_air_in_function_without_lld=
+        [round(pod*10) for pod in pull_out_distance_to_take_transport_air_in_function_without_lld or
+          [5.0]*len(ops)],
+      immersion_depth=[round(id*10) for id in immersion_depth or [0]*len(ops)],
+      surface_following_distance=
+        [round(sfd*10) for sfd in surface_following_distance or [2.1]*len(ops)],
+      tube_2nd_section_height_measured_from_zm=
+        [round(t2sh*10) for t2sh in tube_2nd_section_height_measured_from_zm or [0]*len(ops)],
+      tube_2nd_section_ratio=[round(t2sr*10) for t2sr in tube_2nd_section_ratio or [0]*len(ops)],
+      minimal_traverse_height_at_begin_of_command=
+        [round(mth*10) for mth in
+          minimal_traverse_height_at_begin_of_command or [self._traversal_height]*len(ops)],
+      minimal_height_at_command_end=
+        [round(mh*10) for mh in minimal_height_at_command_end or [self._traversal_height]*len(ops)],
+      dispense_volume=[round(op.volume*100) for op in ops],
+      dispense_speed=[round(fr*10) for fr in flow_rates],
+      cut_off_speed=[round(cs*10) for cs in cut_off_speed or [250]*len(ops)],
+      stop_back_volume=[round(sbv*100) for sbv in stop_back_volume or [0]*len(ops)],
+      transport_air_volume=[round(tav*10) for tav in transport_air_volume or [0]*len(ops)],
+      blow_out_air_volume=[round(boav*100) for boav in blow_out_air_volumes],
+      lld_mode=lld_mode or [0]*len(ops),
+      side_touch_off_distance=round(side_touch_off_distance*10),
+      dispense_position_above_z_touch_off=
+        [round(dpz*10) for dpz in dispense_position_above_z_touch_off or [0.5]*len(ops)],
+      lld_sensitivity=lld_sensitivity or [1]*len(ops),
+      pressure_lld_sensitivity=pressure_lld_sensitivity or [1]*len(ops),
+      swap_speed=[round(ss*10) for ss in swap_speed or [1]*len(ops)],
+      settling_time=[round(st*10) for st in settling_time or [0]*len(ops)],
+      mix_volume=[round(mv*100) for mv in mix_volume or [0]*len(ops)],
+      mix_cycles=mix_cycles or [0]*len(ops),
+      mix_position_in_z_direction_from_liquid_surface=
+        [round(mp) for mp in mix_position_in_z_direction_from_liquid_surface or [0]*len(ops)],
+      mix_speed=[round(ms*10) for ms in mix_speed or [1]*len(ops)],
+      surface_following_distance_during_mixing=
+        [round(sfdm*10) for sfdm in surface_following_distance_during_mixing or [0]*len(ops)],
+      TODO_DD_2=TODO_DD_2 or [0]*len(ops),
       tadm_algorithm_on_off=tadm_algorithm_on_off or 0,
       limit_curve_index=limit_curve_index or [0] * len(ops),
       recording_mode=recording_mode or 0,
@@ -996,8 +899,6 @@ class Vantage(HamiltonLiquidHandler):
   async def aspirate96(
     self,
     aspiration: Union[AspirationPlate, AspirationContainer],
-    jet: bool = False,
-    blow_out: bool = False,
     hlc: Optional[HamiltonLiquidClass] = None,
     type_of_aspiration: int = 0,
     minimal_traverse_height_at_begin_of_command: Optional[float] = None,
@@ -1008,7 +909,6 @@ class Vantage(HamiltonLiquidHandler):
     immersion_depth: float = 0,
     surface_following_distance: float = 0,
     transport_air_volume: Optional[float] = None,
-    blow_out_air_volume: Optional[float] = None,
     pre_wetting_volume: float = 0,
     lld_mode: int = 0,
     lld_sensitivity: int = 4,
@@ -1024,17 +924,8 @@ class Vantage(HamiltonLiquidHandler):
     tadm_algorithm_on_off: int = 0,
     recording_mode: int = 0,
   ):
-    """Aspirate from a plate.
-
-    Args:
-      jet: Whether to find a liquid class with "jet" mode. Only used on dispense.
-      blow_out: Whether to find a liquid class with "blow out" mode. Only used on dispense. Note
-        that this is called "empty" in the VENUS liquid editor, but "blow out" in the firmware
-        documentation.
-      hlc: The Hamiltonian liquid classes to use. If `None`, the liquid classes will be
-        determined automatically based on the tip and liquid used in the first well.
-    """
-    # assert self.core96_head_installed, "96 head must be installed"
+    if hlc is not None:
+      raise NotImplementedError("hlc is deprecated")
 
     if isinstance(aspiration, AspirationPlate):
       top_left_well = aspiration.wells[0]
@@ -1058,35 +949,20 @@ class Vantage(HamiltonLiquidHandler):
 
     liquid_height = position.z + (aspiration.liquid_height or 0)
 
-    tip = aspiration.tips[0]
-    liquid_to_be_aspirated = Liquid.WATER  # default to water
-    if len(aspiration.liquids[0]) > 0 and aspiration.liquids[0][-1][0] is not None:
-      # first part of tuple in last liquid of first well
-      liquid_to_be_aspirated = aspiration.liquids[0][-1][0]
-    if hlc is None:
-      hlc = get_vantage_liquid_class(
-        tip_volume=tip.maximal_volume,
-        is_core=True,
-        is_tip=True,
-        has_filter=tip.has_filter,
-        liquid=liquid_to_be_aspirated,
-        jet=jet,
-        blow_out=blow_out,
-      )
-
-    volume = (
-      hlc.compute_corrected_volume(aspiration.volume) if hlc is not None else aspiration.volume
-    )
-
-    transport_air_volume = transport_air_volume or (
-      hlc.aspiration_air_transport_volume if hlc is not None else 0
-    )
-    blow_out_air_volume = blow_out_air_volume or (
-      hlc.aspiration_blow_out_volume if hlc is not None else 0
-    )
-    flow_rate = aspiration.flow_rate or (hlc.aspiration_flow_rate if hlc is not None else 250)
-    swap_speed = swap_speed or (hlc.aspiration_swap_speed if hlc is not None else 100)
-    settling_time = settling_time or (hlc.aspiration_settling_time if hlc is not None else 5)
+    if transport_air_volume is None:
+      transport_air_volume = 0
+    if aspiration.blow_out_air_volume is None:
+      blow_out_air_volume = 0.0
+    else:
+      blow_out_air_volume = aspiration.blow_out_air_volume
+    if aspiration.flow_rate is None:
+      flow_rate = 250.0
+    else:
+      flow_rate = aspiration.flow_rate
+    if swap_speed is None:
+      swap_speed = 100
+    if settling_time is None:
+      settling_time = 5
 
     return await self.core96_aspiration_of_liquid(
       x_position=round(position.x * 10),
@@ -1104,11 +980,11 @@ class Vantage(HamiltonLiquidHandler):
         pull_out_distance_to_take_transport_air_in_function_without_lld * 10
       ),
       minimum_height=round(well_bottoms * 10),
-      tube_2nd_section_height_measured_from_zm=round(tube_2nd_section_height_measured_from_zm * 10),
-      tube_2nd_section_ratio=round(tube_2nd_section_ratio * 10),
-      immersion_depth=round(immersion_depth * 10),
-      surface_following_distance=round(surface_following_distance * 10),
-      aspiration_volume=round(volume * 100),
+      tube_2nd_section_height_measured_from_zm=round(tube_2nd_section_height_measured_from_zm*10),
+      tube_2nd_section_ratio=round(tube_2nd_section_ratio*10),
+      immersion_depth=round(immersion_depth*10),
+      surface_following_distance=round(surface_following_distance*10),
+      aspiration_volume=round(aspiration.volume * 100),
       aspiration_speed=round(flow_rate * 10),
       transport_air_volume=round(transport_air_volume * 10),
       blow_out_air_volume=round(blow_out_air_volume * 100),
@@ -1150,7 +1026,6 @@ class Vantage(HamiltonLiquidHandler):
     cut_off_speed: float = 250.0,
     stop_back_volume: float = 0,
     transport_air_volume: Optional[float] = None,
-    blow_out_air_volume: Optional[float] = None,
     lld_mode: int = 0,
     lld_sensitivity: int = 4,
     side_touch_off_distance: float = 0,
@@ -1181,6 +1056,9 @@ class Vantage(HamiltonLiquidHandler):
         determined based on the jet, blow_out, and empty parameters.
     """
 
+    if hlc is not None:
+      raise NotImplementedError("hlc is deprecated")
+
     if isinstance(dispense, DispensePlate):
       top_left_well = dispense.wells[0]
       position = (
@@ -1203,36 +1081,24 @@ class Vantage(HamiltonLiquidHandler):
 
     liquid_height = position.z + (dispense.liquid_height or 0) + 10
 
-    tip = dispense.tips[0]
-    liquid_to_be_dispensed = Liquid.WATER  # default to WATER
-    if len(dispense.liquids[0]) > 0 and dispense.liquids[0][-1][0] is not None:
-      # first part of tuple in last liquid of first well
-      liquid_to_be_dispensed = dispense.liquids[0][-1][0]
-    if hlc is None:
-      hlc = get_vantage_liquid_class(
-        tip_volume=tip.maximal_volume,
-        is_core=True,
-        is_tip=True,
-        has_filter=tip.has_filter,
-        liquid=liquid_to_be_dispensed,
-        jet=jet,
-        blow_out=blow_out,  # see method docstring
-      )
-    volume = hlc.compute_corrected_volume(dispense.volume) if hlc is not None else dispense.volume
-
-    transport_air_volume = transport_air_volume or (
-      hlc.dispense_air_transport_volume if hlc is not None else 0
-    )
-    blow_out_air_volume = blow_out_air_volume or (
-      hlc.dispense_blow_out_volume if hlc is not None else 0
-    )
-    flow_rate = dispense.flow_rate or (hlc.dispense_flow_rate if hlc is not None else 250)
-    swap_speed = swap_speed or (hlc.dispense_swap_speed if hlc is not None else 100)
-    settling_time = settling_time or (hlc.dispense_settling_time if hlc is not None else 5)
-    mix_speed = mix_speed or (hlc.dispense_mix_flow_rate if hlc is not None else 100)
-    type_of_dispensing_mode = type_of_dispensing_mode or _get_dispense_mode(
-      jet=jet, empty=empty, blow_out=blow_out
-    )
+    if transport_air_volume is None:
+      transport_air_volume = 0
+    if dispense.blow_out_air_volume is None:
+      blow_out_air_volume = 0.0
+    else:
+      blow_out_air_volume = dispense.blow_out_air_volume
+    if dispense.flow_rate is None:
+      flow_rate = 250.0
+    else:
+      flow_rate = dispense.flow_rate
+    if swap_speed is None:
+      swap_speed = 100
+    if settling_time is None:
+      settling_time = 5
+    if mix_speed is None:
+      mix_speed = 100
+    if type_of_dispensing_mode is None:
+      type_of_dispensing_mode = _get_dispense_mode(jet=jet, empty=empty, blow_out=blow_out)
 
     return await self.core96_dispensing_of_liquid(
       x_position=round(position.x * 10),
@@ -1243,18 +1109,15 @@ class Vantage(HamiltonLiquidHandler):
       tube_2nd_section_ratio=round(tube_2nd_section_ratio * 10),
       lld_search_height=round(lld_search_height * 10),
       liquid_surface_at_function_without_lld=round(liquid_height * 10),
-      pull_out_distance_to_take_transport_air_in_function_without_lld=round(
-        pull_out_distance_to_take_transport_air_in_function_without_lld * 10
-      ),
-      immersion_depth=round(immersion_depth * 10),
-      surface_following_distance=round(surface_following_distance * 10),
-      minimal_traverse_height_at_begin_of_command=round(
-        (minimal_traverse_height_at_begin_of_command or self._traversal_height) * 10
-      ),
-      minimal_height_at_command_end=round(
-        (minimal_height_at_command_end or self._traversal_height) * 10
-      ),
-      dispense_volume=round(volume * 100),
+      pull_out_distance_to_take_transport_air_in_function_without_lld=\
+        round(pull_out_distance_to_take_transport_air_in_function_without_lld*10),
+      immersion_depth=round(immersion_depth*10),
+      surface_following_distance=round(surface_following_distance*10),
+      minimal_traverse_height_at_begin_of_command=
+        round((minimal_traverse_height_at_begin_of_command or self._traversal_height) * 10),
+      minimal_height_at_command_end=
+        round((minimal_height_at_command_end or self._traversal_height)*10),
+      dispense_volume=round(dispense.volume * 100),
       dispense_speed=round(flow_rate * 10),
       cut_off_speed=round(cut_off_speed * 10),
       stop_back_volume=round(stop_back_volume * 100),